+++ src/main/java/com/example/echo/domain/petition/controller/PetitionController.java
@@ -9,6 +9,9 @@ import com.example.echo.domain.petition.dto.response.PetitionDetailResponseDto;
 import com.example.echo.domain.petition.dto.response.PetitionResponseDto;
 import com.example.echo.domain.petition.entity.Category;
 import com.example.echo.domain.petition.service.PetitionService;
+import com.example.echo.global.api.ApiResponse;
+import com.example.echo.global.exception.ErrorCode;
+import com.example.echo.global.exception.PetitionCustomException;
 import com.example.echo.global.security.auth.CustomUserPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -35,50 +38,50 @@ public class PetitionController {
     @PreAuthorize("hasRole('ADMIN')")
     @Operation(summary = "청원 등록", description = "새로운 청원을 등록합니다.")
     @PostMapping
-    public ResponseEntity<PetitionDetailResponseDto> createPetition(@RequestBody PetitionRequestDto petitionDto) {
+    public ResponseEntity<ApiResponse<PetitionDetailResponseDto>> createPetition(@RequestBody PetitionRequestDto petitionDto) {
         PetitionDetailResponseDto createdPetition = petitionService.createPetition(petitionDto);
-        return ResponseEntity.ok(createdPetition);
+        return ResponseEntity.ok(ApiResponse.success(createdPetition));
     }
 
     // 청원 단건 조회
     @Operation(summary = "청원 단건 조회", description = "특정 ID의 청원을 조회합니다.")
     @GetMapping("/{petitionId}")
-    public ResponseEntity<PetitionDetailResponseDto> getPetitionById(@PathVariable Long petitionId) {
+    public ResponseEntity<ApiResponse<PetitionDetailResponseDto>> getPetitionById(@PathVariable Long petitionId) {
         PetitionDetailResponseDto petition = petitionService.getPetitionById(petitionId);
-        return ResponseEntity.ok(petition);
+        return ResponseEntity.ok(ApiResponse.success(petition));
     }
 
     // 청원 전체 조회
     @Operation(summary = "청원 전체 조회", description = "모든 청원을 페이지별로 조회합니다.")
     @GetMapping
-    public ResponseEntity<Page<PetitionResponseDto>> getPetitions(Pageable pageable) {
+    public ResponseEntity<ApiResponse<Page<PetitionResponseDto>>> getPetitions(Pageable pageable) {
         Page<PetitionResponseDto> petitions = petitionService.getOngoingPetitions(pageable);
-        return ResponseEntity.ok(petitions);
+        return ResponseEntity.ok(ApiResponse.success(petitions));
     }
 
     // 청원 카테고리별 조회
     @Operation(summary = "카테고리별 청원 조회", description = "특정 카테고리의 모든 청원을 페이지별로 조회합니다.")
     @GetMapping("/category/{category}")
-    public ResponseEntity<Page<PetitionResponseDto>> getPetitionsByCategory(@PathVariable Category category,
+    public ResponseEntity<ApiResponse<Page<PetitionResponseDto>>> getPetitionsByCategory(@PathVariable Category category,
                                                                             Pageable pageable) {
         Page<PetitionResponseDto> petitions = petitionService.getPetitionsByCategory(pageable, category);
-        return ResponseEntity.ok(petitions);
+        return ResponseEntity.ok(ApiResponse.success(petitions));
     }
 
     // 청원 만료일 순 5개 조회
     @Operation(summary = "청원 만료일 기준 조회", description = "만료일이 가까운 청원 5개를 조회합니다.")
     @GetMapping("/view/endDate")
-    public ResponseEntity<List<PetitionResponseDto>> getEndDatePetitions() {
+    public ResponseEntity<ApiResponse<List<PetitionResponseDto>>> getEndDatePetitions() {
         List<PetitionResponseDto> endDatePetitions = petitionService.getEndDatePetitions();
-        return ResponseEntity.ok(endDatePetitions);
+        return ResponseEntity.ok(ApiResponse.success(endDatePetitions));
     }
 
     // 청원 좋아요 순 5개 조회
     @Operation(summary = "청원 좋아요 수 기준 조회", description = "좋아요 수 많은 청원 5개를 조회합니다.")
     @GetMapping("/view/likesCount")
-    public ResponseEntity<List<PetitionResponseDto>> getLikesCountPetitions() {
+    public ResponseEntity<ApiResponse<List<PetitionResponseDto>>> getLikesCountPetitions() {
         List<PetitionResponseDto> likesCountPetitions = petitionService.getLikesCountPetitions();
-        return ResponseEntity.ok(likesCountPetitions);
+        return ResponseEntity.ok(ApiResponse.success(likesCountPetitions));
     }
 
     // 청원 좋아요 기능
@@ -93,26 +96,26 @@ public class PetitionController {
     // 청원 카테고리 선택 5개 조회
     @Operation(summary = "청원 카테고리별 조회", description = "특정 카테고리의 청원 5개를 랜덤 순으로 조회합니다.")
     @GetMapping("/view/category/{category}")
-    public ResponseEntity<List<PetitionResponseDto>> getRandomCategoryPetitions(@PathVariable Category category) {
+    public ResponseEntity<ApiResponse<List<PetitionResponseDto>>> getRandomCategoryPetitions(@PathVariable Category category) {
         List<PetitionResponseDto> categoryPetitions = petitionService.getRandomCategoryPetitions(category);
-        return ResponseEntity.ok(categoryPetitions);
+        return ResponseEntity.ok(ApiResponse.success(categoryPetitions));
     }
 
     // 제목으로 청원 검색
     @Operation(summary = "청원 제목으로 검색", description = "제목에 검색어가 포함된 청원을 조회합니다.")
     @GetMapping("/search")
-    public ResponseEntity<List<PetitionDetailResponseDto>> searchPetitions(@RequestParam String query) {
+    public ResponseEntity<ApiResponse<List<PetitionDetailResponseDto>>> searchPetitions(@RequestParam String query) {
         List<PetitionDetailResponseDto> petitions = petitionService.searchPetitionsByTitle(query);
-        return ResponseEntity.ok(petitions);
+        return ResponseEntity.ok(ApiResponse.success(petitions));
     }
 
     // 청원 수정
     @PreAuthorize("hasRole('ADMIN')")
     @Operation(summary = "청원 수정", description = "특정 ID의 청원을 수정합니다.")
     @PutMapping("/{petitionId}")
-    public ResponseEntity<PetitionDetailResponseDto> updatePetition(@PathVariable Long petitionId, @RequestBody PetitionRequestDto petitionDto) {
+    public ResponseEntity<ApiResponse<PetitionDetailResponseDto>> updatePetition(@PathVariable Long petitionId, @RequestBody PetitionRequestDto petitionDto) {
         PetitionDetailResponseDto updatedPetition = petitionService.updatePetition(petitionId, petitionDto);
-        return ResponseEntity.ok(updatedPetition);
+        return ResponseEntity.ok(ApiResponse.success(updatedPetition));
     }
 
     // 청원 삭제
@@ -127,7 +130,7 @@ public class PetitionController {
     //관심목록 추가
     @PreAuthorize("authentication.principal.memberId == #requestDTO.memberId")
     @PostMapping("/interestAdd")
-    public ResponseEntity<?> addInterestt(@RequestBody InterestRequestDTO requestDTO) {
+    public ResponseEntity<?> addInterest(@RequestBody InterestRequestDTO requestDTO) {
         try {
             petitionService.addInterest(requestDTO);
             return ResponseEntity.ok("추가되었습니다.");
@@ -158,7 +161,7 @@ public class PetitionController {
     @GetMapping("/Myinterest")
     public ResponseEntity<?> getInterestList(@AuthenticationPrincipal CustomUserPrincipal principal) {
         Member member = memberRepository.findById(principal.getMemberId())
-                .orElseThrow(() -> new EntityNotFoundException("Member not found"));
+                .orElseThrow(() -> new PetitionCustomException(ErrorCode.MEMBER_NOT_FOUND));
         try {
             // 회원의 관심 목록 조회
             List<InterestPetitionResponseDTO> interestList = petitionService.getInterestList(member);
